FIX(appcds): fix building failed on windows
fix issue in SharedClassPathEntry::manifest_size while building on windows.

--story=863510335
Contributed-by:seanwxiang

+++ hotspot/src/share/vm/memory/filemap.hpp
@@ -82,7 +82,7 @@ public:
     return _manifest;
   }
   int manifest_size() const {
-    return (_manifest == NULL) ? 0 : strlen(_manifest);
+    return (_manifest == NULL) ? 0 : (int)strlen(_manifest);
   }
 };
